move geodata handling to version handler
